+++ spec/model/rails_local_analytics_spec.rb
@@ -8,14 +8,6 @@ RSpec.describe RailsLocalAnalytics, type: :model do
   end
 
   context "config.background_jobs" do
-    before do
-      @prev_background_jobs = described_class.config.background_jobs
-    end
-
-    after do
-      described_class.config.background_jobs = @prev_background_jobs
-    end
-
     it "defaults to true" do
       expect(described_class.config.background_jobs).to eq(true)
     end