+++ spec/request/record_request_spec.rb
@@ -26,4 +26,28 @@ RSpec.describe "RailsLocalAnalytics#record_request", type: :request do
       "url_path" => "/",
     })
   end
+
+  it "works with config.background_jobs = false" do
+    RailsLocalAnalytics.config.background_jobs = false
+
+    get root_path
+    expect(response.status).to eq(200)
+
+    expect(TrackedRequestsByDaySite.last.attributes.except("id")).to eq({
+      "browser_engine" => nil,
+      "day" => Date.today,
+      "platform" => nil,
+      "total" => 1,
+      "url_hostname" => "www.example.com",
+    })
+
+    expect(TrackedRequestsByDayPage.last.attributes.except("id")).to eq({
+      "day" => Date.today,
+      "referrer_hostname" => nil,
+      "referrer_path" => nil,
+      "total" => 1,
+      "url_hostname" => "www.example.com",
+      "url_path" => "/",
+    })
+  end
 end